+++ iperf.py
@@ -4,16 +4,16 @@ import perf
 from typing import Any
 from typing import Mapping
 
-import common
+import tftbase
 
 from host import Result
 from logger import logger
 from syncManager import SyncManager
 from testConfig import TestConfig
 from testSettings import TestSettings
-from common import ConnectionMode
-from common import IperfOutput
-from common import TestType
+from tftbase import ConnectionMode
+from tftbase import IperfOutput
+from tftbase import TestType
 from thread import ReturnValueThread
 
 
@@ -36,7 +36,7 @@ class IperfServer(perf.PerfServer):
 
     def setup(self) -> None:
         if self.connection_mode == ConnectionMode.EXTERNAL_IP:
-            cmd = f"podman run -it --rm -p {self.port} --entrypoint {IPERF_EXE} --name={self.pod_name} {common.TFT_TOOLS_IMG} -s --one-off"
+            cmd = f"podman run -it --rm -p {self.port} --entrypoint {IPERF_EXE} --name={self.pod_name} {tftbase.TFT_TOOLS_IMG} -s --one-off"
             cleanup_cmd = f"podman rm --force {self.pod_name}"
         else:
             # Create the server pods
@@ -92,7 +92,7 @@ class IperfClient(perf.PerfClient):
         )
         return json_dump
 
-    def output(self, out: common.TftAggregateOutput) -> None:
+    def output(self, out: tftbase.TftAggregateOutput) -> None:
         # Return machine-readable output to top level
         assert isinstance(
             self._output, IperfOutput
