+++ netperf.py
@@ -1,14 +1,14 @@
 import perf
-import common
+import tftbase
 
 from host import Result
 from logger import logger
 from syncManager import SyncManager
 from testConfig import TestConfig
 from testSettings import TestSettings
-from common import ConnectionMode
-from common import IperfOutput
-from common import TestType
+from tftbase import ConnectionMode
+from tftbase import IperfOutput
+from tftbase import TestType
 from thread import ReturnValueThread
 
 
@@ -28,7 +28,7 @@ class NetPerfServer(perf.PerfServer):
 
     def setup(self) -> None:
         if self.connection_mode == ConnectionMode.EXTERNAL_IP:
-            cmd = f"podman run -it --rm -p {self.port} --entrypoint {NETPERF_SERVER_EXE} --name={self.pod_name} {common.TFT_TOOLS_IMG} -p {self.port} -N"
+            cmd = f"podman run -it --rm -p {self.port} --entrypoint {NETPERF_SERVER_EXE} --name={self.pod_name} {tftbase.TFT_TOOLS_IMG} -p {self.port} -N"
             cleanup_cmd = f"podman rm --force {self.pod_name}"
         else:
             # Create the server pods
@@ -113,7 +113,7 @@ class NetPerfClient(perf.PerfClient):
         )
         return json_dump
 
-    def output(self, out: common.TftAggregateOutput) -> None:
+    def output(self, out: tftbase.TftAggregateOutput) -> None:
         # Return machine-readable output to top level
         assert isinstance(
             self._output, IperfOutput
